aws: improve check for aws_completer - unhardcode path (#9123)
Closes #9122

The current check is assuming that the aws_completer is installed
globally, it then runs the command without that path

I have aws_completer in ~/.local/bin/aws_completer

This changes the check to see if the command exists instead of
concerning itself with where it in installed. This test matches the
behavior of just running the command.